Async JS method invocations
The script host now automatically detects when a JS device initialization, property getter/setter, or method invocation returns a Promise object, and in that case it gets the result via an async callback. The JS bridge is also updated to handle async results from the script host. (The core bridge code already supported async device operations via the IAdapterIoRequest out parameters.)

The BridgeRT and BridgeNet AddDevice APIs are now async and renamed to AddDeviceAsync; that is a build-breaking change to any applications that consume these APIs.

I updated the test device.js script in the WinRT test app, but not in the .NET test app. That makes it easy to test both async and non-async code paths.